rename pydantic_compat module to models
