+++ src/phoenix/server/api/routers/v1/datasets.py
@@ -62,7 +62,7 @@ from phoenix.server.api.types.node import from_global_id_with_expected_type
 from phoenix.server.api.utils import delete_projects, delete_traces
 from phoenix.server.dml_event import DatasetInsertEvent
 
-from .pydantic_compat import V1RoutesBaseModel
+from .models import V1RoutesBaseModel
 from .utils import (
     PaginatedResponseBody,
     ResponseBody,
